fix: Re-organize proerties & replace few control with Style control
