+++ frontend/src/components/BlockPropertySections/OptionsSection.ts
@@ -1,35 +1,11 @@
 import CodeEditor from "@/components/Controls/CodeEditor.vue";
 import InlineInput from "@/components/Controls/InlineInput.vue";
-import OptionToggle from "@/components/Controls/OptionToggle.vue";
 import blockController from "@/utils/blockController";
 
 const setClasses = (val: string) => {
 	const classes = val.split(",").map((c) => c.trim());
 	blockController.setClasses(classes);
 };
-
-const overflowOptions = [
-	{
-		label: "Unset",
-		value: "unset",
-	},
-	{
-		label: "Auto",
-		value: "auto",
-	},
-	{
-		label: "Visible",
-		value: "visible",
-	},
-	{
-		label: "Hidden",
-		value: "hidden",
-	},
-	{
-		label: "Scroll",
-		value: "scroll",
-	},
-];
 
 const optionsSectionProperties = [
 	{
@@ -117,74 +93,6 @@ const optionsSectionProperties = [
 			"update:modelValue": (val: string) => blockController.setKeyValue("innerHTML", val),
 		},
 		condition: () => blockController.isText() || blockController.isButton(),
-	},
-	{
-		component: OptionToggle,
-		getProps: () => {
-			return {
-				label: "Visibility",
-				options: [
-					{
-						label: "Visible",
-						value: "flex",
-					},
-					{
-						label: "Hidden",
-						value: "none",
-					},
-				],
-				modelValue: blockController.getStyle("display") || "flex",
-			};
-		},
-		searchKeyWords: "Visibility, Display, Visible, Hidden, Flex, None, hide, show",
-		events: {
-			"update:modelValue": (val: StyleValue) => blockController.setStyle("display", val),
-		},
-		condition: () => !blockController.isRoot(),
-	},
-	{
-		component: InlineInput,
-		getProps: () => {
-			return {
-				label: "Overflow X",
-				type: "select",
-				options: overflowOptions,
-				modelValue: blockController.getStyle("overflowX", true),
-				placeholder: blockController.getStyle("overflowX", false, true),
-			};
-		},
-		searchKeyWords:
-			"Overflow, X, OverflowX, Overflow X, Auto, Visible, Hide, Scroll, horizontal scroll, horizontalScroll",
-		events: {
-			"update:modelValue": (val: StyleValue) => {
-				if (val === "unset") {
-					val = null;
-				}
-				blockController.setStyle("overflowX", val);
-			},
-		},
-	},
-	{
-		component: InlineInput,
-		getProps: () => {
-			return {
-				label: "Overflow Y",
-				type: "select",
-				options: overflowOptions,
-				modelValue: blockController.getStyle("overflowY", true),
-				placeholder: blockController.getStyle("overflowY", false, true),
-			};
-		},
-		searchKeyWords:
-			"Overflow, Y, OverflowY, Overflow Y, Auto, Visible, Hide, Scroll, vertical scroll, verticalScroll",
-		events: {
-			"update:modelValue": (val: StyleValue) => {
-				if (val === "unset") {
-					val = null;
-				}
-				blockController.setStyle("overflowY", val);
-			},
-		},
 	},
 	{
 		component: InlineInput,
